requirements.txt
Django==1.7.1
Pillow==2.6.1
django-admin-bootstrapped==2.3.1
django-bootstrap3==5.0.3
django-simple-captcha==0.4.3
six==1.8.0
